shared/utility: fixed signedness for comparision
git-svn-id: https://scm.citationtech.net/trunk/shared/utility/utility@1028 b7e6147c-0f76-4c84-a09d-19069afaf10c

+++ utility/map.hpp
@@ -93,7 +93,7 @@ map(const std::string &name, const Sequence &values
 
         SequenceSubRange range
             (values.begin() + start
-             , ((start + length) < values.size())
+             , ((start + length) < static_cast<size_t>(values.size()))
              ? values.begin() + start + length : values.end());
 
         typename ResultList::iterator iresult(result.begin() + start);
